backend/requirements.txt
fastapi
uvicorn
openai
pydub
python-dotenv
notion-client 